# feat(event cache): find related events faster