--- a/crates/matrix-sdk-sqlite/migrations/event_cache_store/009_related_event_index.sql
+++ b/crates/matrix-sdk-sqlite/migrations/event_cache_store/009_related_event_index.sql
@@ -1,0 +1,8 @@
+-- Add an index to speed up queries that look for related events in a room.
+CREATE INDEX "relates_to_idx"
+    ON "events" ("room_id", "relates_to");
+
+-- Add an index to speed up queries that look for related events in a room, with an additional
+-- filter.
+CREATE INDEX "relates_to_rel_type_idx"
+    ON "events" ("room_id", "relates_to", "rel_type");