@bnjbvr 2025-08-26
Could be that the regression when inserting new events is real, since sqlite now has to take care of an extra index. Seems worth it IMO, since writing is still super fast.