@bnjbvr 2025-08-26
Marking ready for review, as this is confirmed fixing the observed issue. I'll update numbers with more up-to-date, now that I've tweaked the benchmark in the other PR.