@poljar 2025-08-26
Semi-related, I wish codspeed would give us some graphical representation for the benchmarks where the performance differs, like criterion gives us.

<img width="6667" height="3750" alt="image" src="https://github.com/user-attachments/assets/f25d0dde-c542-429a-9167-b88225e8693c" />